create teacherDashboard with stack screans
